Undid Multiple Models; Started Triangles
I removed the changes in my last commit to that was to support multiple
models as that is not my focus right now.

I have started the process of changes how I define my models from points
in space to defined triangles.  I created a Triangle class and have
defined a cube using triangles which are made of points.  I also brought
back my Point class to help define triangles.