Almost finished Phong Lighting
I implemented phong's shading equation and I am getting results that are
similar to what I was expecting.  However, I feel there is still a
problem with the position of the light (or maybe the camera) that is
giving a bit wonky lighting effects.  But it looks kind of close.